backend/pyproject.toml
[project]
name = "gez-ai-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.0",
    "firebase-admin>=6.3.0",
    "google-cloud-secret-manager>=2.22.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.27.0",
]
